"Added measurement time calculation in capturarYGestionarDatos function in sensorMIV.cpp"
This commit adds code to calculate and print the time taken to capture 101 readings of temperature and humidity in the capturarYGestionarDatos function in the file sensorMIV.cpp. The time is calculated by storing the current millis() value in two variables, tiempo1 and tiempo2, before and after the readings are taken, and then calculating the difference between the two.